test: refactor and nextRound testing
