+++ plugins/boltz/liquidity.go
@@ -8,6 +8,7 @@ import (
 	"fmt"
 	"sort"
 
+	"github.com/bolt-observer/agent/filter"
 	"github.com/bolt-observer/agent/lightning"
 )
 
@@ -27,24 +28,10 @@ type ChanCapacity struct {
 }
 
 // GetNodeLiquidity - gets current node liquidity
-func (b *Plugin) GetNodeLiquidity(ctx context.Context, optLnAPI lightning.LightingAPICalls) (*Liquidity, error) {
+func GetNodeLiquidity(ctx context.Context, lnAPI lightning.LightingAPICalls, filter filter.FilteringInterface) (*Liquidity, error) {
 	var (
-		lnAPI lightning.LightingAPICalls
-		err   error
+		err error
 	)
-
-	if optLnAPI != nil {
-		lnAPI = optLnAPI
-	} else {
-		lnAPI, err = b.LnAPI()
-		if err != nil {
-			return nil, err
-		}
-		if lnAPI == nil {
-			return nil, fmt.Errorf("could not get lightning api")
-		}
-		defer lnAPI.Cleanup()
-	}
 
 	resp, err := lnAPI.GetChannels(ctx)
 	if err != nil {
@@ -59,7 +46,7 @@ func (b *Plugin) GetNodeLiquidity(ctx context.Context, optLnAPI lightning.Lighti
 	ret.OutboundPercentage = 0.0
 
 	for _, channel := range resp.Channels {
-		if !b.Filter.AllowChanID(channel.ChanID) && !b.Filter.AllowPubKey(channel.RemotePubkey) && !b.Filter.AllowSpecial(channel.Private) {
+		if !filter.AllowChanID(channel.ChanID) && !filter.AllowPubKey(channel.RemotePubkey) && !filter.AllowSpecial(channel.Private) {
 			continue
 		}
 
@@ -77,24 +64,10 @@ func (b *Plugin) GetNodeLiquidity(ctx context.Context, optLnAPI lightning.Lighti
 }
 
 // GetByDescendingOutboundLiquidity - get channels in descreasing outbound liqudity so that sum >= limit satoshis
-func (b *Plugin) GetByDescendingOutboundLiquidity(ctx context.Context, limit uint64, optLnAPI lightning.LightingAPICalls) ([]ChanCapacity, error) {
+func GetByDescendingOutboundLiquidity(ctx context.Context, limit uint64, lnAPI lightning.LightingAPICalls, filter filter.FilteringInterface) ([]ChanCapacity, error) {
 	var (
-		lnAPI lightning.LightingAPICalls
-		err   error
+		err error
 	)
-
-	if optLnAPI != nil {
-		lnAPI = optLnAPI
-	} else {
-		lnAPI, err = b.LnAPI()
-		if err != nil {
-			return nil, err
-		}
-		if lnAPI == nil {
-			return nil, fmt.Errorf("could not get lightning api")
-		}
-		defer lnAPI.Cleanup()
-	}
 
 	resp, err := lnAPI.GetChannels(ctx)
 	if err != nil {
@@ -104,7 +77,7 @@ func (b *Plugin) GetByDescendingOutboundLiquidity(ctx context.Context, limit uin
 	ret := make([]ChanCapacity, 0)
 
 	for _, channel := range resp.Channels {
-		if !b.Filter.AllowChanID(channel.ChanID) && !b.Filter.AllowPubKey(channel.RemotePubkey) && !b.Filter.AllowSpecial(channel.Private) {
+		if !filter.AllowChanID(channel.ChanID) && !filter.AllowPubKey(channel.RemotePubkey) && !filter.AllowSpecial(channel.Private) {
 			continue
 		}
 		// Capacity is outbound liquidity here -> LocalBalance
@@ -139,31 +112,17 @@ func (b *Plugin) GetByDescendingOutboundLiquidity(ctx context.Context, limit uin
 }
 
 // GetChanLiquidity
-func (b *Plugin) GetChanLiquidity(ctx context.Context, chanID uint64, limit uint64, outbound bool, optLnAPI lightning.LightingAPICalls) (*ChanCapacity, uint64, error) {
+func GetChanLiquidity(ctx context.Context, chanID uint64, limit uint64, outbound bool, lnAPI lightning.LightingAPICalls, filter filter.FilteringInterface) (*ChanCapacity, uint64, error) {
 	var (
-		lnAPI lightning.LightingAPICalls
-		err   error
+		err error
 	)
-
-	if optLnAPI != nil {
-		lnAPI = optLnAPI
-	} else {
-		lnAPI, err = b.LnAPI()
-		if err != nil {
-			return nil, 0, err
-		}
-		if lnAPI == nil {
-			return nil, 0, fmt.Errorf("could not get lightning api")
-		}
-		defer lnAPI.Cleanup()
-	}
 
 	resp, err := lnAPI.GetChannels(ctx)
 	if err != nil {
 		return nil, 0, err
 	}
 	for _, channel := range resp.Channels {
-		if !b.Filter.AllowChanID(channel.ChanID) && !b.Filter.AllowPubKey(channel.RemotePubkey) && !b.Filter.AllowSpecial(channel.Private) {
+		if !filter.AllowChanID(channel.ChanID) && !filter.AllowPubKey(channel.RemotePubkey) && !filter.AllowSpecial(channel.Private) {
 			continue
 		}
 		if channel.ChanID != chanID {
